Add Scale trait. Move weight function to fil_cns
