+++ blockchain/chain/Cargo.toml
@@ -14,7 +14,6 @@ forest_crypto   = { version = "0.5", features = ["blst"] }
 encoding        = { package = "forest_encoding", version = "0.2.1" }
 serde           = { version = "1.0", features = ["derive"] }
 forest_bigint   = "0.1"
-num-traits      = "0.2"
 forest_message  = { version = "0.7", features = ["blst"] }
 ipld_blockstore = "0.1"
 thiserror       = "1.0"
