+++ blockchain/consensus/fil_cns/Cargo.toml
@@ -5,12 +5,14 @@ authors = ["ChainSafe Systems <info@chainsafe.io>"]
 edition = "2021"
 
 [dependencies]
+anyhow = "1.0"
 async-std = { version = "1.9", features = ["tokio1", "unstable"] }
 async-trait = "0.1"
 futures = "0.3.5"
 log = "0.4.8"
 nonempty = "0.8.0"
 thiserror = "1.0"
+num-traits = "0.2"
 
 address = { package = "forest_address", version = "0.3" }
 cid = { package = "forest_cid", version = "0.3" }
@@ -18,6 +20,8 @@ crypto = { package = "forest_crypto", version = "0.5", features = ["blst"] }
 encoding = { package = "forest_encoding", version = "0.2.1" }
 ipld_blockstore = "0.1"
 fil_types = "0.2"
+fvm = "1.0"
+fvm_shared      = { version = "0.8.0", default-features = false }
 
 actor = { package = "actor_interface", path = "../../../vm/actor_interface" }
 beacon = { path = "../../beacon" }
